Started Initial Backend MyData Work + CSS Theme Change
-Started backend preparations for the new MyData pages, but still needs
tweaking and talking with the team to get the relevant data presented to
the user. Because of this, the next page is not accessible by normal
means
-The CSS theme has been changed to more closely match the website, with
a blue-white gradient in the background and a silver general theme that
looks much better than the old MyRelevate blue.
-Also removed the buttons at the top-right, it just seemed to make more
sense to put everything in the slide-menu for usability and readability.